Re-integrate openai api
